bug修复

+++ app/src/main/java/com/marvin/splashedinkkotlin/utils/glide/MyGlideModule.kt
@@ -21,7 +21,7 @@ class MyGlideModule : GlideModule {
 
     override fun applyOptions(context: Context?, builder: GlideBuilder?) {
         if (builder != null) {
-            builder.setDiskCache(DiskLruCacheFactory(BuildConfig.image_cache, DEFAULT_DISK_CACHE_SIZE))
+            builder.setDiskCache(DiskLruCacheFactory(BuildConfig.image_cache, DEFAULT_DISK_CACHE_SIZE.toLong()))
         }
     }
 }
